remove configparser class from rtfits -> create specific file - use pathlib instead of os.path in configparser - add possibility to programmatically import modules - add possibility to provide additional processing-arguments
